formats/elf/reloc: do not use .rela./.rel. but use dynamic entries (#1643)
In executable/lib ELFs the loader only use the info provided in the
PT_DYNAMIC segment. There are entries that specify the
size/entsize/address of the relocation tables and those should be used
instead of relying on sections info.

Sections can be trusted only in ET_REL elfs, where the section are the
only kind of information we have.

This new test adds a ELF executable that has been manually changed to
shrink the .rela.dyn section. If we rely on sections, we miss some
relocation entries that should instead be available if the binary needs
to run correctly.