Rewrite the RBuffer API to make it safer and adjust the codebase ##refactor (#1828)
With the new RBuffer implementation, the non-read parts of the buffer
are completed with 0xff_priv, which is 0 by default. This changes a bit
how the binary is read, but it better matches how the loader works and
how the process sees the memory, as mapped pages usually contain 0 by
default and not 0xff.